Move the "migrations" file to "database/core" folder.
